requirements[1].txt
discord.py
aiosqlite
matplotlib
python-dotenv
requests
aiohttp
humanize
pytz